+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v4.0.0
+New features:
+- Add RACE processor.
+- Add option for custom log filtering.
+Changes:
+- Add microsecond delay unit.
+
 ## v3.0.0
 New features:
 - Optional multithreaded filter processing.